refactor: move safely_get_field method to firebase_loader
The safely_get_field function, responsible for retrieving values safely from a Firestore document snapshot, has been moved from app_config.py to firebase_loader.py. This change included removing redundant import statements and adapting function calls to reflect the new location.
